Switch to local settings, better fix for WebView bug
Switched to storing preferences on the local machine so that they
persist between sessions. We use the Settings VM to help with that and
x:Bind to it.

New fix for a58e07ab6a3d5b86cd192e0225affb986ac80ad0. It turns out that
if you Task.Delay AT ALL in BackRequested, Windows will think you
pressed and held the back button when you're in tablet mode.

For whatever reason, setting NavigationCacheMode="Enabled" on
WebPage.xaml fixes the bug (seems like WebView only likes to be
constructed once). This has the side-effect of showing the last viewed
page before navigating to the new one. For example, if you went to
reddit.com, then back, then to google.com, the user would see a flash of
reddit.com before google.com starts loading.

To fix this, we add landingpage.html to our package and always navigate
to it first. It's blank, so the user doesn't see a flash of anything.
Why is it a file in our package and not just a string? WebView only
seems to add to its history if you navigate to an actual file. So the
new order of events is: user wants to navigate to reddit.com. We first
go to landingpage.html, then wait 300ms, then reddit.com. User clicks
back, landingpage.html is loaded. We detect that case and call
Frame.GoBack() for them. User now wants to go to google.com. We first go
to landingpage.html, wait 300ms, then go to google.com.